test: Split shape-related tests into their own module
